moved init.js from Proxyapp > Proxyapp/Cursor Added Proxyapp/QRLWallet Added init.js to Proxyapp/QRLWallet Added Readme.md to Proxyapp/QRLWallet Updated main Readme.md to Add QRLWallet to known vulnerable list
